requirements.txt
flask
flask-cors
streamlit
google-generativeai
python-dotenv
requests
mcp
